SonarCloud workaround
