+++ server/auth/authDB.js
@@ -1,9 +1,11 @@
+'use strict';
+
 const database = require('../database/database.js');
 const auth = require('./auth.js');
 const bcrypt = require('bcrypt');
 const saltRounds = 10;
 
-database.getUserLogin = (/* { params } */) => {
+const getUserLogin = (/* { params } */) => {
   return new Promise((resolve, reject) => {
     resolve();
     // db.getUserLogin(reviewIdFilter)
@@ -17,7 +19,7 @@ database.getUserLogin = (/* { params } */) => {
 };
 
 
-database.getUser = (user) => {
+const getUser = (user) => {
   return new Promise((resolve, reject) => {
     database.User.find({ username: user })
       .then((res) => {
@@ -35,7 +37,32 @@ database.getUser = (user) => {
   });
 };
 
-database.login = (username, password) => {
+const verifySession = (token) => {
+  console.log('Attempting to verify session', token);
+  return new Promise((resolve, reject) => {
+    const sessionData = auth.session(token);
+    // console.log(sessionData);
+    if (!sessionData) {
+      resolve('expired');
+    }
+    database.User.find({ username: sessionData.username })
+      .then(response => {
+        const session = response[0].session;
+        // console.log(session);
+        if (session !== token) {
+          resolve('invalid');
+        } else {
+          resolve(session);
+        }
+      })
+      .catch(err => {
+        console.log('authService:', err);
+        reject(err);
+      });
+  });
+};
+
+const login = (username, password) => {
   return new Promise((resolve, reject) => {
     database.User.find({ username: username })
       .then((res) => {
@@ -44,7 +71,7 @@ database.login = (username, password) => {
             throw 'Incorrect password';
           }
           const sessionToken = auth.createSessionToken(username);
-          database.updateUser({ username: username }, { session: sessionToken })
+          updateUser({ username: username }, { session: sessionToken })
             .then(result => {
               console.log(`Session record created for ${username}:`, sessionToken);
               resolve(sessionToken);
@@ -63,7 +90,7 @@ database.login = (username, password) => {
 };
 
 
-database.addUser = (userObj) => {
+const addUser = (userObj) => {
   return new Promise((resolve, reject) => {
     const filter = { $or: [{ username: userObj.username }, { email: userObj.email }] };
     bcrypt.hash(userObj.password, saltRounds, function (err, hash) {
@@ -88,7 +115,7 @@ database.addUser = (userObj) => {
 };
 
 
-database.updateUser = (filter, update) => {
+const updateUser = (filter, update) => {
   // console.log(filter, update);
   return new Promise((resolve, reject) => {
     database.User.findOneAndUpdate(filter, update, { upsert: true })
@@ -103,4 +130,4 @@ database.updateUser = (filter, update) => {
   });
 };
 
-module.exports.database = database;
+module.exports = { login, getUser, addUser, updateUser, verifySession };
